Move useTokenLoansTransactions to TokenLoansContent
