+++ src/pages/tokenLending/LoansTokenPage/LoansTokenActiveTable/hooks/index.ts
@@ -1,2 +1,1 @@
 export * from './useLoansTokenActiveTable'
-export * from './useTokenLoansTransactions'
