+++ src/pages/tokenLending/LoansTokenPage/TokenLoansContent/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useTokenLoansContent'
+export * from './useTokenLoansTransactions'
