+++ src/pages/tokenLending/LoansTokenPage/TokenLoansContent/hooks/useTokenLoansTransactions.ts
@@ -3,7 +3,7 @@ import { every, uniqueId } from 'lodash'
 import moment from 'moment'
 import { TxnExecutor } from 'solana-transactions-executor'
 
-import { core } from '@banx/api/tokens'
+import { TokenLoan } from '@banx/api/tokens'
 import { useIsLedger, useModal } from '@banx/store/common'
 import { useTokenLoansOptimistic } from '@banx/store/token'
 import {
@@ -31,8 +31,11 @@ import {
   isTokenLoanRepaymentCallActive,
 } from '@banx/utils'
 
-import { caclFractionToRepay, caclFractionToRepayForRepaymentCall } from '../helpers'
-import { useSelectedTokenLoans } from '../loansState'
+import {
+  caclFractionToRepay,
+  caclFractionToRepayForRepaymentCall,
+} from '../../LoansTokenActiveTable/helpers'
+import { useSelectedTokenLoans } from '../loansCart'
 
 export const useTokenLoansTransactions = () => {
   const wallet = useWallet()
@@ -44,7 +47,7 @@ export const useTokenLoansTransactions = () => {
   const { update: updateLoansOptimistic } = useTokenLoansOptimistic()
   const { selection, clear: clearSelection } = useSelectedTokenLoans()
 
-  const repayLoan = async (loan: core.TokenLoan) => {
+  const repayLoan = async (loan: TokenLoan) => {
     const loadingSnackbarId = uniqueId()
 
     try {
@@ -83,14 +86,11 @@ export const useTokenLoansTransactions = () => {
               const { bondTradeTransaction, fraktBond } =
                 parseRepayLoanSimulatedAccounts(accountInfoByPubkey)
 
-              //TODO Move optimistics creation into a separate function
-              const optimisticLoan: core.TokenLoan = {
-                ...params.loan,
-                publicKey: fraktBond.publicKey,
-                //? Needs to prevent BE data overlap in optimistics logic
-                fraktBond: { ...fraktBond, lastTransactedAt: moment().unix() },
-                bondTradeTransaction: bondTradeTransaction,
-              }
+              const optimisticLoan = createOptimisticLoan(
+                params.loan,
+                fraktBond,
+                bondTradeTransaction,
+              )
 
               updateLoansOptimistic([optimisticLoan], wallet.publicKey.toBase58())
               clearSelection()
@@ -146,14 +146,11 @@ export const useTokenLoansTransactions = () => {
                 const { bondTradeTransaction, fraktBond } =
                   parseRepayLoanSimulatedAccounts(accountInfoByPubkey)
 
-                //TODO Move optimistics creation into a separate function
-                const optimisticLoan: core.TokenLoan = {
-                  ...params.loan,
-                  publicKey: fraktBond.publicKey,
-                  //? Needs to prevent BE data overlap in optimistics logic
-                  fraktBond: { ...fraktBond, lastTransactedAt: moment().unix() },
-                  bondTradeTransaction: bondTradeTransaction,
-                }
+                const optimisticLoan = createOptimisticLoan(
+                  params.loan,
+                  fraktBond,
+                  bondTradeTransaction,
+                )
 
                 updateLoansOptimistic([optimisticLoan], wallet.publicKey.toBase58())
               }
@@ -181,7 +178,7 @@ export const useTokenLoansTransactions = () => {
     }
   }
 
-  const repayPartialLoan = async (loan: core.TokenLoan, fractionToRepay: number) => {
+  const repayPartialLoan = async (loan: TokenLoan, fractionToRepay: number) => {
     const loadingSnackbarId = uniqueId()
 
     try {
@@ -223,14 +220,11 @@ export const useTokenLoansTransactions = () => {
               const { bondTradeTransaction, fraktBond } =
                 parseRepayPartialLoanSimulatedAccounts(accountInfoByPubkey)
 
-              //TODO Move optimistics creation into a separate function
-              const optimisticLoan: core.TokenLoan = {
-                ...params.loan,
-                publicKey: fraktBond.publicKey,
-                //? Needs to prevent BE data overlap in optimistics logic
-                fraktBond: { ...fraktBond, lastTransactedAt: moment().unix() },
-                bondTradeTransaction: bondTradeTransaction,
-              }
+              const optimisticLoan = createOptimisticLoan(
+                params.loan,
+                fraktBond,
+                bondTradeTransaction,
+              )
 
               updateLoansOptimistic([optimisticLoan], wallet.publicKey.toBase58())
               clearSelection()
@@ -304,14 +298,11 @@ export const useTokenLoansTransactions = () => {
                 const { bondTradeTransaction, fraktBond } =
                   parseRepayPartialLoanSimulatedAccounts(accountInfoByPubkey)
 
-                //TODO Move optimistics creation into a separate function
-                const optimisticLoan: core.TokenLoan = {
-                  ...params.loan,
-                  publicKey: fraktBond.publicKey,
-                  //? Needs to prevent BE data overlap in optimistics logic
-                  fraktBond: { ...fraktBond, lastTransactedAt: moment().unix() },
-                  bondTradeTransaction: bondTradeTransaction,
-                }
+                const optimisticLoan = createOptimisticLoan(
+                  params.loan,
+                  fraktBond,
+                  bondTradeTransaction,
+                )
 
                 updateLoansOptimistic([optimisticLoan], wallet.publicKey.toBase58())
               }
@@ -345,5 +336,23 @@ export const useTokenLoansTransactions = () => {
     repayAllLoans,
     repayPartialLoan,
     repayUnpaidLoansInterest,
+  }
+}
+
+const createOptimisticLoan = (
+  loan: TokenLoan,
+  newFraktBond: TokenLoan['fraktBond'],
+  newBondTradeTransaction: TokenLoan['bondTradeTransaction'],
+): TokenLoan => {
+  const currentTimeInSeconds = moment().unix()
+
+  return {
+    ...loan,
+    publicKey: newFraktBond.publicKey,
+    fraktBond: {
+      ...newFraktBond,
+      lastTransactedAt: currentTimeInSeconds, //? Needs to prevent BE data overlap in optimistics logic
+    },
+    bondTradeTransaction: newBondTradeTransaction,
   }
 }
